Bug 888789 - Update nestegg with VP9 and Opus support. r=kinetik
Update our in-tree source to the upstream master branch.

UltraBlame original commit: 28f89a6c0b7b99caf15763d0c1d9415b90e8c549

+++ media/libnestegg/src/macros.h
@@ -20,7 +20,7 @@
 
 
 
-#define structof(p,t,f) ((t*)(- offsetof(t,f) + (char*)(p)))
+#define structof(p,t,f) ((t*)(- (ptrdiff_t) offsetof(t,f) + (char*)(p)))
